test: reorganize worker test cases
