+++ tests/integration/worker/constructor/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "worker-constructor-test",
+  "version": "1.0.0",
+  "private": true,
+  "type": "module"
+}
