+++ tests/integration/worker/constructor/rslib.config.ts
