+++ tests/integration/worker/constructor/src/helper.ts
