+++ tests/integration/worker/constructor/src/index.ts
