+++ tests/integration/worker/constructor/src/worker.ts
