+++ tests/integration/worker/query-inline/package.json
