+++ tests/integration/worker/query-inline/rslib.config.ts
