+++ tests/integration/worker/query-inline/src/index.ts
