+++ tests/integration/worker/query-inline/src/worker.ts
