+++ tests/integration/worker/query/package.json
