+++ tests/integration/worker/query/src/index.ts
