+++ tests/integration/worker/query/src/worker.ts
